Change honeypot to check for key presence.
Browsers appear to do a pretty good job of not adding form inputs that
are `display:none` to the request at all. Some bots however, will
include the honeypot with no value.

Previously, this would pass the `invisible_captcha?` check, but could
fail in the controller for people using
`ActionController::StrongParameters`, which fails on presence of key in
the `params` hash. Now `invisible_captcha?` triggers if the honeypot
exists as a key in `params`.