Add examples for `invisible_captcha.spam_detected` event
Initially I was testing that ActiveSupport::Notifications received `instrument` with the expected arguments, but ultimately decided that approach was testing the *implementation* rather than the intended effect.

So for this test, we create a dummy event subscriber and assert that it is invoked with the expected arguments when spam is detected.